Fix: move dynamic import to client component (ssr:false not allowed in Server Components)
