+++ src/app/reading-room/voice/page.tsx
@@ -1,14 +1,5 @@
 import type { Metadata } from 'next';
-import dynamic from 'next/dynamic';
-
-const VoiceAgentClient = dynamic(() => import('./VoiceAgentClient'), {
-  ssr: false,
-  loading: () => (
-    <div className="min-h-screen bg-stone-50 flex items-center justify-center">
-      <p className="text-stone-400">Loading voice agent...</p>
-    </div>
-  ),
-});
+import VoiceAgentLoader from './VoiceAgentLoader';
 
 export const metadata: Metadata = {
   title: 'Voice Research — The Reading Room — Source Library',
@@ -16,5 +7,5 @@ export const metadata: Metadata = {
 };
 
 export default function VoiceAgentPage() {
-  return <VoiceAgentClient />;
+  return <VoiceAgentLoader />;
 }
